fixing bug with terraform state transition for bastion
for some reason when you add a bastion to the host the first run of
terraform apply doesn't see the bastion, but see the changes in ssh
settings (port in my case).
found this while writing post about the provider (applied test
configuration to localhost).
teraform sdk is a piece of crap, it is simplier to just make all
settings flat than try to maintain some sort of hierarchy
(which is more obvious because bastion uses ssh.config if it is defined)